requirements.txt
Django==6.0.3
Pillow==12.1.1
gunicorn==23.0.0
djangorestframework==3.16.1
djangorestframework-simplejwt==5.4.0
requests~=2.32.3
